A number of fixes and changes to facilitate smooth integration.
 - Added proxy initialization when not using a mock proxy.
 - Changed marshaling for space-padded strings to be NULL-character
   agnostic and thus more tolerant of applications which do not strictly
   adhere to the PKCS #11 spec regarding the format of these strings.
 - In ChapsServiceRedirect: delay initialization of the target library
   to the first method call and change effective permissions to
   "chronos:pkcs11" before initializing.
 - Grabbed the latest common.mk from cros-disks.

BUG=chromium-os:21005
TEST=make runtests
     manual test on x86-zgb
     > ./chapsd_test --use_dbus

+++ chaps/chaps.h
@@ -9,6 +9,6 @@ namespace chaps {
 
 extern const char* kChapsServicePath;
 extern const char* kChapsServiceName;
-extern const int kTokenLabelSize;
+extern const size_t kTokenLabelSize;
 
 }  // namespace
